✨ Add: new animation `scale-up-origin-top`
